+++ js/randomizer.js
@@ -1,0 +1,6 @@
+const createRandomInteger = (min, max) => {
+  const result = Math.random() * (min - max + 1) + max;
+  return Math.floor(result);
+};
+
+export {createRandomInteger};
